scripts/check_neon_connection.py
"""Load DATABASE_URL from .env and verify Neon/Postgres connectivity."""

from __future__ import annotations

import os
import sys
from pathlib import Path

from dotenv import load_dotenv
import psycopg


def main() -> int:
    repo_root = Path(__file__).resolve().parents[1]
    load_dotenv(repo_root / ".env")
    url = os.environ.get("DATABASE_URL")
    if not url:
        print("DATABASE_URL is not set. Copy .env.example to .env and add your URL.", file=sys.stderr)
        return 1

    with psycopg.connect(url, connect_timeout=15) as conn:
        with conn.cursor() as cur:
            cur.execute("SELECT 1 AS ok")
            row = cur.fetchone()
    print("Connection OK:", row)
    return 0


if __name__ == "__main__":
    raise SystemExit(main())
