requirements.txt
psycopg[binary]>=3.2,<4
python-dotenv>=1.0,<2
dbt-postgres>=1.8,<2
